Update forms to use FormComponent
