+++ ui/packages/ui/src/app/shared/index.ts
@@ -1,5 +1,4 @@
 export * from './Utils';
-export * from './CategoryUtil';
 export * from './WithErrorBoundary';
 export * from './ApiError';
 export * from './ConfirmationDialog';
